// src/shared/dataprovider/api/model/loginRequestDTO.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface LoginRequestDTO { 
    username?: string;
    password?: string;
}