// src/shared/dataprovider/api/model/loginResultDTO.ts
/**
 * Dynasty Cash In Gateway
 * Dynasty Cash In API list
 *
 * OpenAPI spec version: v1
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Client } from './client';
import { ShortenedAgentUser } from './shortenedAgentUser';

export interface LoginResultDTO { 
    Id?: number;
    Username?: string;
    Name?: string;
    Token : string;
    RefreshToken?: string;
    PasswordDidChanged?: boolean;
    Client?: Client;
    Agent?: ShortenedAgentUser;
}